pyproject.toml
[tool.poetry]
name = "scraper"
version = "0.1.0"
description = "a little fastapi parser for AI mobile app"
authors = ["Hoaper <beeboplay@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.3"
lxml-html-clean = "^0.1.1"
websocket = "^0.2.1"
selenium = "^4.21.0"
fastapi-cache2 = "^0.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
